+++ CHANGELOG.md
@@ -27,8 +27,9 @@ Resource data will be available with the resource object. This enhancement helps
 - Connection template
 - Enclosure
 - FC network
-- Interconnect type
 - SAS interconnect type
+- SAS logical interconnect group
+- Interconnect type
 
 # 4.8.0
 #### Notes