+++ CHANGELOG.md
@@ -24,10 +24,12 @@ Resource data will be available with the resource object. This enhancement helps
   Refer example files of the updated resources for more details.
 
 #### features supported with current release
+- Connection template
 - Enclosure
-- FC Network
+- FC network
+- Interconnect type
 
-# 4.8.0 (Unreleased)
+# 4.8.0
 #### Notes
 Added the capability to handle OneView Appliance SNMP Settings
 Extends support of the SDK to OneView Rest API version 800 (OneView v4.1).